Basic working networked play, but without transmission of data from guest to computer.

+++ src/main/java/com/socialthingy/qaopm/spectrum/remote/SpectrumState.java
@@ -3,17 +3,17 @@ package com.socialthingy.qaopm.spectrum.remote;
 import java.io.Serializable;
 
 public class SpectrumState implements Serializable {
-    private byte[] screen;
+    private int[] screen;
     private int[] borderLines;
     private boolean flashActive;
 
-    public SpectrumState(final byte[] screen, final int[] borderLines, final boolean flashActive) {
+    public SpectrumState(final int[] screen, final int[] borderLines, final boolean flashActive) {
         this.screen = screen;
         this.borderLines = borderLines;
         this.flashActive = flashActive;
     }
 
-    public byte[] getScreen() {
+    public int[] getScreen() {
         return screen;
     }
